todo_app/phase_2/backend/pyproject.toml
[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Backend for Todo App Phase 2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlmodel",
    "pyjwt",
    "python-dotenv",
    "psycopg2-binary",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
]

[build-system]

requires = ["hatchling"]

build-backend = "hatchling.build"



[tool.hatch.build.targets.wheel]

packages = ["src/todo_app"]
